refactor(report): one class per output format behind ReportRendererInterface
Replace the @internal ReportRenderer god class (renderConsole/renderJson/
renderSarif/renderHtml/renderMarkdown + the sibling JunitReportRenderer) with
six focused renderers each implementing the new ReportRendererInterface
(format(): string + render(AuditReport): string): ConsoleReportRenderer,
JsonReportRenderer, SarifReportRenderer, HtmlReportRenderer,
MarkdownReportRenderer, JunitReportRenderer.

Shared concerns move to ReportPackage (name/homepage/version) and TemplateLoader
(template reads). Renderers are autoconfigured via a ReportRendererInterface
instanceof tag (symfony_security_auditor.report_renderer); ReportWriter takes the
tagged iterator, indexes by format() key, and dispatches the selected --format,
throwing UnsupportedOutputFormatException when no renderer advertises the key.

All six --format values and their output are unchanged; ReportRenderer was
@internal so its removal is not a BC break. Adding a format is now "new class +
service registration" with no match arm to edit. Split ReportRendererTest into
per-renderer tests over a shared AbstractReportRendererTestCase.
