+++ CLAUDE.md
@@ -89,7 +89,7 @@ src/
       Pricing/       # StaticPricingProvider
       Progress/      # ConsoleProgressReporter (decorated TTY), PlainProgressReporter (CI/non-TTY), LoggerProgressReporter, ProgressReporterHolder, ProgressContext, AuditOverviewLine
       Tool/          # ReadFileTool, GrepTool, ListFilesTool, LookupAdvisoryTool, SymfonyToolRegistryFactory, RecordVulnerabilityTool, RecordVulnerabilityToolFactory, RecordReviewTool, RecordReviewToolFactory
-      Report/        # ReportRenderer (console/json/sarif/html; + Template/*.txt + *.html stubs)
+      Report/        # ReportRendererInterface (format/render) + one class per format ({Console,Json,Sarif,Html,Markdown,Junit}ReportRenderer) + ReportPackage + TemplateLoader; + Template/*.txt + *.html stubs
   Command/           # AuditCommand (Symfony Console: audit:run, alias audit) + AuditCommandInput, AuditPresenter, ReportWriter, AuditExitCodeResolver, ExitCode enum, AuditCommandHelp, OutputFormat enum (console|json|sarif|html), Baseline (accepted-finding suppression)
   SymfonySecurityAuditorBundle.php  # Bundle class (configure + loadExtension)
 tests/Phpunit/
