+++ docs/extending.md
@@ -450,38 +450,45 @@ $report->toArray(): array<string, mixed>             // fully serializable; incl
 
 ### Built-in formats
 
-`ReportRenderer` ships six formats out of the box:
+Every format is its own class implementing `ReportRendererInterface`
+(`format(): string` + `render(AuditReport): string`). One class per format keeps
+each renderer small and independently testable. The bundle ships six:
 
-- `renderConsole(AuditReport): string` — human-readable terminal output
+- `ConsoleReportRenderer` (`console`) — human-readable terminal output
   (templates in `Report/Template/*.txt`).
-- `renderJson(AuditReport): string` — pretty-printed `AuditReport::toArray()`.
-- `renderSarif(AuditReport): string` — SARIF 2.1.0, consumable by GitHub Code
+- `JsonReportRenderer` (`json`) — pretty-printed `AuditReport::toArray()`.
+- `SarifReportRenderer` (`sarif`) — SARIF 2.1.0, consumable by GitHub Code
   Scanning and GitLab Security Dashboard.
-- `renderHtml(AuditReport): string` — self-contained HTML page (templates in
+- `HtmlReportRenderer` (`html`) — self-contained HTML page (templates in
   `Report/Template/*.html`).
-- `JunitReportRenderer::render(AuditReport): string` — JUnit XML, one failed
-  test case per finding, for CI test-report panels (e.g. GitLab merge-request
-  widgets on every tier). Lives in its own class so `ReportRenderer` does not
-  keep growing — prefer this shape for new formats.
-- `renderMarkdown(AuditReport): string` — Markdown suitable for PR comments and
+- `MarkdownReportRenderer` (`markdown`) — Markdown suitable for PR comments and
   wikis.
+- `JunitReportRenderer` (`junit`) — JUnit XML, one failed test case per finding,
+  for CI test-report panels (e.g. GitLab merge-request widgets on every tier).
+
+Each renderer is tagged `symfony_security_auditor.report_renderer` (via the
+`ReportRendererInterface` `instanceof` autoconfiguration in
+`config/services.php`). `Command\ReportWriter` receives the tagged iterator,
+indexes the renderers by their `format()` key, and dispatches the selected
+`--format` to the matching one — throwing `UnsupportedOutputFormatException` if
+no renderer advertises that key.
 
 Trigger them via `audit:run --format=console|json|sarif|html|markdown|junit`
 (see [`ci.md`](ci.md) for SARIF upload workflows).
 
 ### Adding a new format
 
-1. Add a case to `Command\OutputFormat`.
-2. Add a dedicated `<Name>ReportRenderer` class with a
-   `render(AuditReport): string` method (see `JunitReportRenderer` for the
-   shape) — or, for a small variation on an existing format, a
-   `render<Name>(AuditReport): string` method on `ReportRenderer`.
-3. Add the matching arm to the `match` block in `Command\ReportWriter::write()`.
+1. Add a case to `Command\OutputFormat` with the wire value.
+2. Add a `<Name>ReportRenderer` class implementing `ReportRendererInterface`,
+   returning that same wire value from `format()` and building the output in
+   `render()` (see any bundled renderer for the shape).
+3. Register the service in `config/services.php` — autoconfiguration tags it,
+   and `ReportWriter` picks it up automatically; no `match` arm to edit.
 
 `AuditReport` is a plain value object — once the pipeline completes you may also
 inject it directly into any consumer (custom command, controller, event
 listener) and serialize it however fits your output target without going through
-`ReportRenderer`.
+a renderer.
 
 ## 5. Other Pluggable Ports
