+++ src/Audit/Infrastructure/Report/MarkdownReportRenderer.php
@@ -1,0 +1,106 @@
+<?php
+
+/*
+ * This file is part of the vinceamstoutz/symfony-security-auditor package.
+ *
+ * (c) Vincent Amstoutz <vincent.amstoutz.dev@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace VinceAmstoutz\SymfonySecurityAuditor\Audit\Infrastructure\Report;
+
+use Override;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Domain\Model\AuditReport;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Domain\Model\Vulnerability;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Domain\Model\VulnerabilitySeverity;
+
+/** @internal not part of the BC promise — see docs/versioning.md */
+final readonly class MarkdownReportRenderer implements ReportRendererInterface
+{
+    #[Override]
+    public function format(): string
+    {
+        return 'markdown';
+    }
+
+    #[Override]
+    public function render(AuditReport $auditReport): string
+    {
+        return implode("\n", [
+            '# Security Audit Report',
+            '',
+            \sprintf(
+                '**Risk level:** %s (score %d) · **Findings:** %d · **Files scanned:** %d',
+                $auditReport->riskLevel(),
+                $auditReport->riskScore(),
+                $auditReport->totalVulnerabilities(),
+                $auditReport->filesScanned(),
+            ),
+            '',
+            $this->body($auditReport),
+            '',
+            '---',
+            '',
+            \sprintf('Generated by [%s](%s).', ReportPackage::NAME, ReportPackage::HOMEPAGE_URL),
+        ]);
+    }
+
+    private function body(AuditReport $auditReport): string
+    {
+        if (0 === $auditReport->totalVulnerabilities()) {
+            return '✅ No validated vulnerabilities found.';
+        }
+
+        $lines = ['## Summary by severity', '', '| Severity | Count |', '| --- | --- |'];
+
+        foreach (VulnerabilitySeverity::cases() as $severity) {
+            $count = \count($auditReport->vulnerabilitiesBySeverity($severity));
+            if ($count > 0) {
+                $lines[] = \sprintf('| %s | %d |', $severity->label(), $count);
+            }
+        }
+
+        $lines[] = '';
+        $lines[] = \sprintf('## Findings (%d)', $auditReport->totalVulnerabilities());
+
+        foreach ($auditReport->vulnerabilities() as $vulnerability) {
+            $lines[] = '';
+            $lines[] = $this->vulnerability($vulnerability);
+        }
+
+        return implode("\n", $lines);
+    }
+
+    private function vulnerability(Vulnerability $vulnerability): string
+    {
+        return implode("\n", [
+            \sprintf('### %s — %s', $vulnerability->severity()->label(), $vulnerability->title()),
+            '',
+            \sprintf('- **Type:** `%s` (%s)', $vulnerability->type()->value, $vulnerability->type()->owaspReference()),
+            \sprintf('- **Location:** `%s:%d-%d`', $vulnerability->filePath(), $vulnerability->lineStart(), $vulnerability->lineEnd()),
+            \sprintf('- **Confidence:** %s%%', \sprintf('%.0f', $vulnerability->confidence() * 100)),
+            '',
+            $vulnerability->description(),
+            '',
+            \sprintf('**Attack vector:** %s', $vulnerability->attackVector()),
+            '',
+            '**Proof:**',
+            '',
+            $this->codeBlock($vulnerability->proof()),
+            '',
+            \sprintf('**Remediation:** %s', $vulnerability->remediation()),
+        ]);
+    }
+
+    private function codeBlock(string $text): string
+    {
+        return implode("\n", array_map(
+            static fn (string $line): string => \sprintf('    %s', $line),
+            explode("\n", $text),
+        ));
+    }
+}
