+++ src/Command/Exception/UnsupportedOutputFormatException.php
@@ -1,0 +1,25 @@
+<?php
+
+/*
+ * This file is part of the vinceamstoutz/symfony-security-auditor package.
+ *
+ * (c) Vincent Amstoutz <vincent.amstoutz.dev@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace VinceAmstoutz\SymfonySecurityAuditor\Command\Exception;
+
+use RuntimeException;
+
+/** @internal not part of the BC promise — see docs/versioning.md */
+final class UnsupportedOutputFormatException extends RuntimeException
+{
+    public static function forFormat(string $format): self
+    {
+        return new self(\sprintf('No report renderer is registered for output format "%s".', $format));
+    }
+}
