+++ tests/Phpunit/Unit/Command/ReportWriterTest.php
@@ -21,7 +21,13 @@ use Symfony\Component\Console\Style\SymfonyStyle;
 use Symfony\Component\Filesystem\Filesystem;
 use VinceAmstoutz\SymfonySecurityAuditor\Audit\Domain\Model\AuditContext;
 use VinceAmstoutz\SymfonySecurityAuditor\Audit\Domain\Model\AuditReport;
-use VinceAmstoutz\SymfonySecurityAuditor\Audit\Infrastructure\Report\ReportRenderer;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Infrastructure\Report\ConsoleReportRenderer;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Infrastructure\Report\HtmlReportRenderer;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Infrastructure\Report\JsonReportRenderer;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Infrastructure\Report\JunitReportRenderer;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Infrastructure\Report\MarkdownReportRenderer;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Infrastructure\Report\SarifReportRenderer;
+use VinceAmstoutz\SymfonySecurityAuditor\Command\Exception\UnsupportedOutputFormatException;
 use VinceAmstoutz\SymfonySecurityAuditor\Command\OutputFormat;
 use VinceAmstoutz\SymfonySecurityAuditor\Command\ReportWriter;
 
@@ -37,7 +43,14 @@ final class ReportWriterTest extends TestCase
     protected function setUp(): void
     {
         $this->filesystem = new Filesystem();
-        $this->reportWriter = new ReportWriter(new ReportRenderer(), $this->filesystem);
+        $this->reportWriter = new ReportWriter([
+            new ConsoleReportRenderer(),
+            new JsonReportRenderer(),
+            new SarifReportRenderer(),
+            new HtmlReportRenderer(),
+            new MarkdownReportRenderer(),
+            new JunitReportRenderer(),
+        ], $this->filesystem);
         $this->tmpDir = sys_get_temp_dir().'/report_writer_test_'.uniqid('', true);
         $this->filesystem->mkdir($this->tmpDir);
     }
@@ -48,6 +61,9 @@ final class ReportWriterTest extends TestCase
         $this->filesystem->remove($this->tmpDir);
     }
 
+    /**
+     * @throws UnsupportedOutputFormatException
+     */
     public function test_writing_to_file_announces_save_path_via_success_message(): void
     {
         $bufferedOutput = new BufferedOutput();
@@ -62,6 +78,9 @@ final class ReportWriterTest extends TestCase
         self::assertStringContainsString($outputFile, $display);
     }
 
+    /**
+     * @throws UnsupportedOutputFormatException
+     */
     public function test_writing_to_file_persists_content_to_disk(): void
     {
         $symfonyStyle = new SymfonyStyle(new StringInput(''), new BufferedOutput());
@@ -75,6 +94,9 @@ final class ReportWriterTest extends TestCase
         self::assertIsArray(json_decode($content, true));
     }
 
+    /**
+     * @throws UnsupportedOutputFormatException
+     */
     public function test_writing_without_file_streams_content_to_console(): void
     {
         $bufferedOutput = new BufferedOutput();
@@ -87,6 +109,9 @@ final class ReportWriterTest extends TestCase
         self::assertStringNotContainsString('Report saved to', $display);
     }
 
+    /**
+     * @throws UnsupportedOutputFormatException
+     */
     public function test_writing_html_format_streams_an_html_document(): void
     {
         $bufferedOutput = new BufferedOutput();
@@ -99,6 +124,9 @@ final class ReportWriterTest extends TestCase
         self::assertStringContainsString('Security Audit Report', $display);
     }
 
+    /**
+     * @throws UnsupportedOutputFormatException
+     */
     public function test_writing_junit_format_streams_a_junit_xml_document(): void
     {
         $bufferedOutput = new BufferedOutput();
@@ -111,6 +139,9 @@ final class ReportWriterTest extends TestCase
         self::assertStringContainsString('symfony-security-auditor', $display);
     }
 
+    /**
+     * @throws UnsupportedOutputFormatException
+     */
     public function test_writing_to_file_creates_missing_parent_directories(): void
     {
         $symfonyStyle = new SymfonyStyle(new StringInput(''), new BufferedOutput());
@@ -119,6 +150,20 @@ final class ReportWriterTest extends TestCase
         $this->reportWriter->write($this->makeReport(), OutputFormat::Json, $outputFile, $symfonyStyle);
 
         self::assertFileExists($outputFile);
+    }
+
+    /**
+     * @throws UnsupportedOutputFormatException
+     */
+    public function test_writing_a_format_without_a_registered_renderer_throws(): void
+    {
+        $reportWriter = new ReportWriter([new JsonReportRenderer()], $this->filesystem);
+        $symfonyStyle = new SymfonyStyle(new StringInput(''), new BufferedOutput());
+
+        $this->expectException(UnsupportedOutputFormatException::class);
+        $this->expectExceptionMessage('No report renderer is registered for output format "sarif".');
+
+        $reportWriter->write($this->makeReport(), OutputFormat::Sarif, null, $symfonyStyle);
     }
 
     private function makeReport(): AuditReport
