+++ tests/Phpunit/Unit/Infrastructure/Report/HtmlReportRendererTest.php
@@ -1,0 +1,199 @@
+<?php
+
+/*
+ * This file is part of the vinceamstoutz/symfony-security-auditor package.
+ *
+ * (c) Vincent Amstoutz <vincent.amstoutz.dev@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace VinceAmstoutz\SymfonySecurityAuditor\Tests\Unit\Infrastructure\Report;
+
+use Override;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Domain\Exception\InvalidCodeLocationException;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Domain\Exception\InvalidVulnerabilityClassificationException;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Domain\Model\AuditCost;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Domain\Model\CodeLocation;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Domain\Model\Vulnerability;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Domain\Model\VulnerabilityClassification;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Domain\Model\VulnerabilityNarrative;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Domain\Model\VulnerabilitySeverity;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Domain\Model\VulnerabilityType;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Infrastructure\Report\HtmlReportRenderer;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Infrastructure\Report\ReportRendererInterface;
+
+final class HtmlReportRendererTest extends AbstractReportRendererTestCase
+{
+    #[Override]
+    protected function createRenderer(): ReportRendererInterface
+    {
+        return new HtmlReportRenderer();
+    }
+
+    public function test_it_advertises_the_html_format(): void
+    {
+        self::assertSame('html', $this->renderer->format());
+    }
+
+    public function test_render_is_a_complete_document(): void
+    {
+        $output = $this->renderer->render($this->makeReport());
+
+        self::assertStringContainsString('<!doctype html>', $output);
+        self::assertStringContainsString('</html>', $output);
+    }
+
+    public function test_render_footer_links_to_the_project_homepage(): void
+    {
+        $output = $this->renderer->render($this->makeReport());
+
+        self::assertStringContainsString(
+            'Generated by <a href="https://github.com/vinceamstoutz/symfony-security-auditor">vinceamstoutz/symfony-security-auditor</a>.',
+            $output,
+        );
+    }
+
+    public function test_render_shows_safe_message_when_no_vulnerabilities(): void
+    {
+        $output = $this->renderer->render($this->makeReport());
+
+        self::assertStringContainsString('No validated vulnerabilities found.', $output);
+        self::assertStringNotContainsString('<h2>Vulnerabilities', $output);
+    }
+
+    public function test_render_renders_risk_level_with_its_lowercased_css_class(): void
+    {
+        $output = $this->renderer->render($this->makeReport());
+
+        self::assertStringContainsString('class="risk risk-safe"', $output);
+        self::assertStringContainsString('>SAFE</span>', $output);
+    }
+
+    public function test_render_shows_the_primary_model(): void
+    {
+        $output = $this->renderer->render($this->makeReportWithCost(AuditCost::zero('claude-test-model')));
+
+        self::assertStringContainsString('claude-test-model', $output);
+        self::assertStringNotContainsString('unknown model', $output);
+    }
+
+    public function test_render_shows_unknown_model_when_the_primary_model_is_blank(): void
+    {
+        $output = $this->renderer->render($this->makeReportWithCost(AuditCost::zero('')));
+
+        self::assertStringContainsString('unknown model', $output);
+    }
+
+    /**
+     * @throws InvalidCodeLocationException
+     * @throws InvalidVulnerabilityClassificationException
+     */
+    public function test_render_lists_a_validated_vulnerability(): void
+    {
+        $output = $this->renderer->render($this->makeReport($this->makeValidatedVuln(filePath: 'src/Admin/UserController.php')));
+
+        self::assertStringContainsString('<h2>Vulnerabilities (1 total)</h2>', $output);
+        self::assertStringContainsString('src/Admin/UserController.php', $output);
+        self::assertStringContainsString('class="finding severity-high"', $output);
+    }
+
+    /**
+     * @throws InvalidCodeLocationException
+     * @throws InvalidVulnerabilityClassificationException
+     */
+    public function test_render_summary_counts_only_present_severities(): void
+    {
+        $output = $this->renderer->render($this->makeReport($this->makeValidatedVuln(vulnerabilitySeverity: VulnerabilitySeverity::HIGH)));
+
+        self::assertStringContainsString('<tr class="severity-high">', $output);
+        self::assertStringContainsString('<td>1</td></tr>', $output);
+        self::assertStringNotContainsString('<tr class="severity-critical">', $output);
+    }
+
+    /**
+     * @throws InvalidCodeLocationException
+     * @throws InvalidVulnerabilityClassificationException
+     */
+    public function test_render_escapes_finding_content_to_prevent_report_xss(): void
+    {
+        $vulnerability = Vulnerability::of(
+            new VulnerabilityClassification(VulnerabilityType::SQL_INJECTION, VulnerabilitySeverity::HIGH, '<script>alert(1)</script>', 0.9),
+            new CodeLocation('src/Foo.php', 1, 2),
+            new VulnerabilityNarrative('desc', 'vec', 'proof', 'fix'),
+            'code',
+        )->withReviewerValidation(true);
+
+        $output = $this->renderer->render($this->makeReport($vulnerability));
+
+        self::assertStringContainsString('&lt;script&gt;alert(1)&lt;/script&gt;', $output);
+        self::assertStringNotContainsString('<script>alert(1)</script>', $output);
+    }
+
+    /**
+     * @throws InvalidCodeLocationException
+     * @throws InvalidVulnerabilityClassificationException
+     */
+    public function test_render_renders_confidence_as_a_percentage(): void
+    {
+        $output = $this->renderer->render($this->makeReport($this->makeValidatedVuln()));
+
+        self::assertStringContainsString('90%', $output);
+    }
+
+    /**
+     * @throws InvalidCodeLocationException
+     * @throws InvalidVulnerabilityClassificationException
+     */
+    public function test_render_replaces_every_template_placeholder(): void
+    {
+        $output = $this->renderer->render($this->makeReport($this->makeValidatedVuln()));
+
+        self::assertStringNotContainsString('{{', $output);
+    }
+
+    /**
+     * @throws InvalidCodeLocationException
+     * @throws InvalidVulnerabilityClassificationException
+     */
+    public function test_render_entity_encodes_single_quotes_via_ent_quotes(): void
+    {
+        $output = $this->renderer->render($this->makeReport($this->makeValidatedVuln()));
+
+        self::assertStringContainsString('&#039; OR 1=1', $output);
+        self::assertStringNotContainsString("<pre>' OR 1=1", $output);
+    }
+
+    /**
+     * @throws InvalidCodeLocationException
+     * @throws InvalidVulnerabilityClassificationException
+     */
+    public function test_render_renders_the_location_with_file_and_line_range(): void
+    {
+        $output = $this->renderer->render($this->makeReport(
+            $this->makeValidatedVuln(filePath: 'src/Repo.php', lineStart: 10),
+        ));
+
+        self::assertStringContainsString('<dd>src/Repo.php:10-14</dd>', $output);
+    }
+
+    /**
+     * @throws InvalidCodeLocationException
+     * @throws InvalidVulnerabilityClassificationException
+     */
+    public function test_render_summary_table_renders_exactly_one_row_per_present_severity(): void
+    {
+        $output = $this->renderer->render($this->makeReport(
+            $this->makeValidatedVuln(vulnerabilitySeverity: VulnerabilitySeverity::HIGH),
+        ));
+
+        $expectedTable = '<table class="summary"><caption>Summary by severity</caption>'
+            .'<tr class="severity-high"><th>'.VulnerabilitySeverity::HIGH->label().'</th><td>1</td></tr>'
+            .'</table>';
+
+        self::assertStringContainsString($expectedTable, $output);
+    }
+}
