+++ tests/Phpunit/Unit/Infrastructure/Report/MarkdownReportRendererTest.php
@@ -1,0 +1,127 @@
+<?php
+
+/*
+ * This file is part of the vinceamstoutz/symfony-security-auditor package.
+ *
+ * (c) Vincent Amstoutz <vincent.amstoutz.dev@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace VinceAmstoutz\SymfonySecurityAuditor\Tests\Unit\Infrastructure\Report;
+
+use Override;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Domain\Exception\InvalidCodeLocationException;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Domain\Exception\InvalidVulnerabilityClassificationException;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Domain\Model\VulnerabilitySeverity;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Infrastructure\Report\MarkdownReportRenderer;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Infrastructure\Report\ReportRendererInterface;
+
+final class MarkdownReportRendererTest extends AbstractReportRendererTestCase
+{
+    #[Override]
+    protected function createRenderer(): ReportRendererInterface
+    {
+        return new MarkdownReportRenderer();
+    }
+
+    public function test_it_advertises_the_markdown_format(): void
+    {
+        self::assertSame('markdown', $this->renderer->format());
+    }
+
+    public function test_render_starts_with_the_report_heading(): void
+    {
+        $output = $this->renderer->render($this->makeReport());
+
+        self::assertStringContainsString('# Security Audit Report', $output);
+    }
+
+    public function test_render_footer_links_to_the_project_homepage(): void
+    {
+        $output = $this->renderer->render($this->makeReport());
+
+        self::assertStringContainsString(
+            'Generated by [vinceamstoutz/symfony-security-auditor](https://github.com/vinceamstoutz/symfony-security-auditor).',
+            $output,
+        );
+    }
+
+    public function test_render_shows_safe_message_when_no_vulnerabilities(): void
+    {
+        $output = $this->renderer->render($this->makeReport());
+
+        self::assertStringContainsString('No validated vulnerabilities found.', $output);
+        self::assertStringNotContainsString('## Findings', $output);
+    }
+
+    public function test_render_includes_the_risk_level(): void
+    {
+        $output = $this->renderer->render($this->makeReport());
+
+        self::assertStringContainsString('SAFE', $output);
+    }
+
+    /**
+     * @throws InvalidCodeLocationException
+     * @throws InvalidVulnerabilityClassificationException
+     */
+    public function test_render_lists_a_validated_finding_with_its_location(): void
+    {
+        $output = $this->renderer->render($this->makeReport(
+            $this->makeValidatedVuln(filePath: 'src/Admin/UserController.php', lineStart: 10),
+        ));
+
+        self::assertStringContainsString('## Findings', $output);
+        self::assertStringContainsString('Test Vuln', $output);
+        self::assertStringContainsString('`src/Admin/UserController.php:10-14`', $output);
+    }
+
+    /**
+     * @throws InvalidCodeLocationException
+     * @throws InvalidVulnerabilityClassificationException
+     */
+    public function test_render_finding_shows_confidence_percent_and_indented_proof(): void
+    {
+        $output = $this->renderer->render($this->makeReport($this->makeValidatedVuln()));
+
+        self::assertStringContainsString('**Confidence:** 90%', $output);
+        self::assertStringContainsString("\n    ' OR 1=1", $output);
+    }
+
+    /**
+     * @throws InvalidCodeLocationException
+     * @throws InvalidVulnerabilityClassificationException
+     */
+    public function test_render_renders_a_severity_summary_table(): void
+    {
+        $output = $this->renderer->render($this->makeReport(
+            $this->makeValidatedVuln(vulnerabilitySeverity: VulnerabilitySeverity::HIGH),
+        ));
+
+        self::assertStringContainsString("## Summary by severity\n\n| Severity | Count |\n| --- | --- |", $output);
+        self::assertStringContainsString(\sprintf('| %s | 1 |', VulnerabilitySeverity::HIGH->label()), $output);
+        self::assertStringNotContainsString(VulnerabilitySeverity::CRITICAL->label(), $output);
+    }
+
+    /**
+     * @throws InvalidCodeLocationException
+     * @throws InvalidVulnerabilityClassificationException
+     */
+    public function test_render_orders_findings_most_severe_first(): void
+    {
+        $vulnerability = $this->makeValidatedVuln(vulnerabilitySeverity: VulnerabilitySeverity::LOW, filePath: 'src/Low.php');
+        $critical = $this->makeValidatedVuln(vulnerabilitySeverity: VulnerabilitySeverity::CRITICAL, filePath: 'src/Critical.php');
+
+        $output = $this->renderer->render($this->makeReport($vulnerability, $critical));
+
+        $criticalPosition = strpos($output, 'src/Critical.php');
+        $lowPosition = strpos($output, 'src/Low.php');
+        self::assertNotFalse($criticalPosition);
+        self::assertNotFalse($lowPosition);
+        self::assertLessThan($lowPosition, $criticalPosition);
+    }
+}
